+++ crates/client/src/client/schema/tx.rs
@@ -51,7 +51,19 @@ pub struct TxIdArgs {
 )]
 pub struct TransactionQuery {
     #[arguments(id: $id)]
-    pub transaction: Option<OpaqueTransaction>,
+    pub transaction: Option<OpaqueTransactionWithStatus>,
+}
+
+/// Retrieves the transaction in opaque form
+#[derive(cynic::QueryFragment, Clone, Debug)]
+#[cynic(
+    schema_path = "./assets/schema.sdl",
+    graphql_type = "Query",
+    variables = "TxIdArgs"
+)]
+pub struct TransactionStatusQuery {
+    #[arguments(id: $id)]
+    pub transaction: Option<OpaqueTransactionStatus>,
 }
 
 #[derive(cynic::QueryFragment, Clone, Debug)]
@@ -92,13 +104,25 @@ impl TryFrom<TransactionConnection> for PaginatedResult<TransactionResponse, Str
 #[cynic(schema_path = "./assets/schema.sdl")]
 pub struct TransactionEdge {
     pub cursor: String,
-    pub node: OpaqueTransaction,
+    pub node: OpaqueTransactionWithStatus,
 }
 
 #[derive(cynic::QueryFragment, Clone, Debug)]
 #[cynic(graphql_type = "Transaction", schema_path = "./assets/schema.sdl")]
 pub struct OpaqueTransaction {
     pub raw_payload: HexString,
+}
+
+#[derive(cynic::QueryFragment, Clone, Debug)]
+#[cynic(graphql_type = "Transaction", schema_path = "./assets/schema.sdl")]
+pub struct OpaqueTransactionWithStatus {
+    pub raw_payload: HexString,
+    pub status: Option<TransactionStatus>,
+}
+
+#[derive(cynic::QueryFragment, Clone, Debug)]
+#[cynic(graphql_type = "Transaction", schema_path = "./assets/schema.sdl")]
+pub struct OpaqueTransactionStatus {
     pub status: Option<TransactionStatus>,
 }
 
@@ -178,7 +202,8 @@ pub struct SubmittedStatus {
 #[derive(cynic::QueryFragment, Clone, Debug)]
 #[cynic(schema_path = "./assets/schema.sdl")]
 pub struct SuccessStatus {
-    pub transaction_id: TransactionId,
+    #[cfg(feature = "test-helpers")]
+    pub transaction: OpaqueTransaction,
     pub block_height: U32,
     pub time: Tai64Timestamp,
     pub program_state: Option<ProgramState>,
@@ -190,7 +215,8 @@ pub struct SuccessStatus {
 #[derive(cynic::QueryFragment, Clone, Debug)]
 #[cynic(schema_path = "./assets/schema.sdl")]
 pub struct FailureStatus {
-    pub transaction_id: TransactionId,
+    #[cfg(feature = "test-helpers")]
+    pub transaction: OpaqueTransaction,
     pub block_height: U32,
     pub time: Tai64Timestamp,
     pub reason: String,
@@ -418,6 +444,7 @@ pub mod tests {
     use crate::client::schema::Bytes;
     use fuel_core_types::fuel_types::canonical::Serialize;
 
+    #[cfg(not(feature = "test-helpers"))]
     #[test]
     fn transparent_transaction_by_id_query_gql_output() {
         use cynic::QueryBuilder;
@@ -427,6 +454,7 @@ pub mod tests {
         insta::assert_snapshot!(operation.query)
     }
 
+    #[cfg(not(feature = "test-helpers"))]
     #[test]
     fn opaque_transaction_by_id_query_gql_output() {
         use cynic::QueryBuilder;
@@ -436,6 +464,7 @@ pub mod tests {
         insta::assert_snapshot!(operation.query)
     }
 
+    #[cfg(not(feature = "test-helpers"))]
     #[test]
     fn transactions_connection_query_gql_output() {
         use cynic::QueryBuilder;
@@ -448,6 +477,7 @@ pub mod tests {
         insta::assert_snapshot!(operation.query)
     }
 
+    #[cfg(not(feature = "test-helpers"))]
     #[test]
     fn transactions_by_owner_gql_output() {
         use cynic::QueryBuilder;